refactor: Improve acts detail page structure and maintainability
- Extract helper functions to lib/utils.ts for better reusability
- Create lib/data.ts for centralized data fetching logic
- Simplify layout.tsx by extracting complex business logic
- Reduce sidebar.tsx from 402 to 346 lines by removing inline functions
- Consolidate mobile-scroll.css with cleaner structure
- Improve type safety and remove unused exports
- Maintain same UI/UX while significantly improving code organization
